+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [10.0.0] - 2026-02-08
+### Added
+* Add support for Symfony 8.0
+
+### Changed
+* Update to PHPUnit 13
+
+### Deprecated
+* *Nothing*
+
+### Removed
+* Remove QR code config options that were deprecated.
+* Remove append extra path config option that was deprecated.
+* Drop support for PHP 8.3.
+* Drop support for Symfony 7.x.
+
+### Fixed
+* *Nothing*
+
+
 ## [9.7.0] - 2025-11-01
 ### Added
 * Add `REDIS_SERVERS_USER` and `REDIS_SERVERS_PASSWORD` config options.